Rename test packages, Add controller tests for pagination
