+++ src/test/java/com/bayou/managers/UnverifiedUserManagerTests.java
@@ -1,4 +1,4 @@
-package com.bayou.TestManagers;
+package com.bayou.managers;
 
 import com.bayou.MainConfig;
 import com.bayou.managers.impl.UnverifiedUserManager;
